+++ packages/util/src/bn/toHex.ts
@@ -7,8 +7,6 @@ import type { BN } from './bn';
 import { u8aToHex } from '../u8a';
 import { bnToU8a } from './toU8a';
 
-const ZERO_STR = '0x00';
-
 /**
  * @name bnToHex
  * @summary Creates a hex value from a BN.js bignumber object.
@@ -25,7 +23,5 @@ const ZERO_STR = '0x00';
  * ```
  */
 export function bnToHex <ExtToBn extends ToBn> (value?: ExtToBn | BN | bigint | number | null, { bitLength = -1, isLe = false, isNegative = false }: NumberOptions = {}): HexString {
-  return !value
-    ? ZERO_STR
-    : u8aToHex(bnToU8a(value, { bitLength, isLe, isNegative }));
+  return u8aToHex(bnToU8a(value, { bitLength, isLe, isNegative }));
 }